Added new page for Add new life group
