+++ nlfyadminconsole/package/src/routes/Router.js
@@ -12,8 +12,11 @@ const Alerts = lazy(() => import("../views/ui/Alerts"));
 const Announcements = lazy(() => import("../views/ui/Announcements"));
 const ChurchPrayers = lazy(() => import("../views/ui/ChurchPrayers"));
 const PrayerRequests = lazy(() => import("../views/ui/PrayerRequests"));
-const LifeGroups = lazy(() => import("../views/ui/LifeGroups"));
-const LifeGroupDetails = lazy(() => import("../views/ui/LifeGroupDetails"));
+const AddLifeGroup = lazy(() => import("../views/ui/LifeGroups/AddLifeGroup"));
+const LifeGroups = lazy(() => import("../views/ui/LifeGroups/LifeGroups"));
+const LifeGroupDetails = lazy(() =>
+  import("../views/ui/LifeGroups/LifeGroupDetails")
+);
 // const Test = lazy(() => import("../views/ui/test"));
 const Badges = lazy(() => import("../views/ui/Badges"));
 const Buttons = lazy(() => import("../views/ui/Buttons"));
@@ -38,6 +41,7 @@ const ThemeRoutes = [
       { path: "/churchPrayers", exact: true, element: <ChurchPrayers /> },
       { path: "/prayerRequests", exact: true, element: <PrayerRequests /> },
       { path: "/lifeGroups", exact: true, element: <LifeGroups /> },
+      { path: "/addLifeGroups", exact: true, element: <AddLifeGroup /> },
       // { path: "/test", exact: true, element: <Test /> },
       {
         path: "/lifeGroupsDetails/:id",
